+++ src/main/resources/db/migration/V2__INSERT INTO TABLES.sql
@@ -1,0 +1,10 @@
+INSERT INTO USERS VALUES (1,'Ciobanu Alin','alin.m.ciobanu@gmail.com',1);
+INSERT INTO USERS VALUES (2,'Testing Migration','testingmigration@gmail.com',2);
+INSERT INTO BANK VALUES (1,1000);
+INSERT INTO BANK VALUES (2,2000);
+INSERT INTO PRODUCTS(product_id,product_name,product_price,product_quantity) VALUES(1,'cana',20,100);
+INSERT INTO PRODUCTS(product_id,product_name,product_price,product_quantity) VALUES(2,'vaza',80,20);
+INSERT INTO PRODUCTS(product_id,product_name,product_price,product_quantity) VALUES(3,'pian',5000,3);
+INSERT INTO TYPES VALUES(1,'vesela');
+INSERT INTO TYPES VALUES(2,'decoratiuni');
+INSERT INTO TYPES VALUES(3,'muzica');